requirements.txt
pytesseract
pdf2image
Pillow
pandas
google-generativeai
gradio
